pyproject.toml
[tool.poetry]
name = "Allure telegram bot"
version = "1.0.0"
description = "Telegram bot for testing management"
authors = ["Eftech Team"]

[tool.poetry.dependencies]
python = ">=3.9 <3.10"
emoji = "^1.7.0"
matplotlib = "^3.5.1"
pytz = "^2022.1"
pymongo = "^4.1.0"
python-dotenv = "^0.20.0"
python-gitlab = "^3.3.0"
python-telegram-bot = "^13.11"
requests = "^2.27.1"
click = "^8.1.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
mccabe = "0.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
